Moved `HashWriter` into its own module.
